Roles y permisos

+++ app/Http/Controllers/RolesController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
+
 
 use Spatie\Permission\Models\Role;
 
@@ -14,6 +16,47 @@ class RolesController extends Controller
         $roles = roles()->all();
 
         return view('admin.roles.index', compact('roles'));
+    }
+
+    public function modify(Request $request, $id)
+    {
+        $role = Role::findOrFail($id);
+
+        return view ('admin.roles.update', compact('role'));
+    }
+
+    public function update(Request $request, $id)
+    {
+        $role = Role::findOrFail($id);
+
+        # validaciones.
+        $validator = \Validator::make($request->all(), [
+            'name' => 'required',
+            'slug' => [
+                'required',
+                Rule::unique((new \App\Role)->getTable(), 'name')
+                    ->ignore($role->name, 'name')
+                    ->ignore($role->id)
+            ],
+            'route' => 'required'
+        ]);
+
+        if ($validator->fails()) {
+            return redirect()
+                        ->route('roles.modify', $role->id)
+                        ->withErrors($validator)
+                        ->withInput();
+        }
+
+        # store
+        $role = roles($role)->update([
+            'name' => $request->name, 
+            'slug' => $request->slug, 
+            'route' => $request->route
+        ]);
+
+        # return
+        return redirect()->route('roles.index')->with('action', 'updated');
     }
     
     public function store(Request $request, $id = null)
@@ -48,10 +91,14 @@ class RolesController extends Controller
 
     public function destroy(Request $request, $id)
     {
-        Role::destroy($id);
+        if($id > 2) {
+            Role::destroy($id);
+            
+            return redirect()
+                    ->route('roles.index')
+                    ->with('action', 'delete');
+        } 
 
-        return redirect()
-                ->route('roles.index')
-                ->with('action', delete);
+        return redirct()->route('roles.index');
     }
 }
